refactor: useProductManager 훅 생성
