+++ src/refactoring/components/Admin/index.ts
@@ -1,0 +1,3 @@
+export * from "./AdminItem";
+export * from "./ManageCouponForm";
+export * from "./NewProductForm";
