+++ src/refactoring/hooks/index.ts
@@ -1,3 +1,4 @@
 export * from "./useCart.ts";
 export * from "./useCoupon.ts";
 export * from "./useProduct.ts";
+export * from "./useProductManager.ts";
