+++ src/refactoring/views/AdminPage.tsx
@@ -1,10 +1,12 @@
-import { useState } from "react";
-import { Coupon, Discount, Product } from "../../types.ts";
+import { Coupon, Product } from "../../types.ts";
 import { TitleMenu } from "../components/Common/TitleMenu.tsx";
 import { Button } from "../components/Common/Button.tsx";
-import { NewProductForm } from "../components/Admin/NewProductForm.tsx";
-import { AdminItem } from "../components/Admin/AdminItem.tsx";
-import { ManageCouponForm } from "../components/Admin/ManageCouponForm.tsx";
+import {
+  AdminItem,
+  ManageCouponForm,
+  NewProductForm,
+} from "../components/Admin";
+import { useProductManager } from "../hooks";
 interface Props {
   products: Product[];
   coupons: Coupon[];
@@ -20,58 +22,23 @@ export const AdminPage = ({
   onProductAdd,
   onCouponAdd,
 }: Props) => {
-  const [openProductIds, setOpenProductIds] = useState<Set<string>>(new Set());
-  const [editingProduct, setEditingProduct] = useState<Product | null>(null);
-  const [newDiscount, setNewDiscount] = useState<Discount>({
-    quantity: 0,
-    rate: 0,
-  });
-  const [newCoupon, setNewCoupon] = useState<Coupon>({
-    name: "",
-    code: "",
-    discountType: "percentage",
-    discountValue: 0,
-  });
-  const [showNewProductForm, setShowNewProductForm] = useState(false);
-  const [newProduct, setNewProduct] = useState<Omit<Product, "id">>({
-    name: "",
-    price: 0,
-    stock: 0,
-    discounts: [],
-  });
-  // custom hook으로 빼기
-  const toggleProductAccordion = (productId: string) => {
-    setOpenProductIds((prev) => {
-      const newSet = new Set(prev);
-      if (newSet.has(productId)) {
-        newSet.delete(productId);
-      } else {
-        newSet.add(productId);
-      }
-      return newSet;
-    });
-  };
-
-  // handleEditProduct 함수 수정
-  const handleEditProduct = (product: Product) => {
-    setEditingProduct({ ...product });
-  };
-
-  // 새로운 핸들러 함수 추가
-  const handleProductNameUpdate = (productId: string, newName: string) => {
-    if (editingProduct && editingProduct.id === productId) {
-      const updatedProduct = { ...editingProduct, name: newName };
-      setEditingProduct(updatedProduct);
-    }
-  };
-
-  // 새로운 핸들러 함수 추가
-  const handlePriceUpdate = (productId: string, newPrice: number) => {
-    if (editingProduct && editingProduct.id === productId) {
-      const updatedProduct = { ...editingProduct, price: newPrice };
-      setEditingProduct(updatedProduct);
-    }
-  };
+  const {
+    openProductIds,
+    editingProduct,
+    newDiscount,
+    newCoupon,
+    showNewProductForm,
+    newProduct,
+    setEditingProduct,
+    setNewDiscount,
+    setNewCoupon,
+    setShowNewProductForm,
+    toggleProductAccordion,
+    setNewProduct,
+    handleEditProduct,
+    handleProductNameUpdate,
+    handlePriceUpdate,
+  } = useProductManager();
 
   // 수정 완료 핸들러 함수 추가
   const handleEditComplete = () => {
